requirements/requirements_docs.txt
Sphinx==4.4.0
jupyter_sphinx==0.3.2
numpydoc==1.2
pyansys_sphinx_theme==0.2.3
pypandoc==1.7.2
pytest-sphinx==0.3.1
sphinx-autobuild==2021.3.14
sphinx-autodoc-typehints==1.17.0
sphinx-copybutton==0.5.0
sphinx-gallery==0.10.1
sphinx-notfound-page==0.8
sphinxcontrib-websupport==1.2.4
sphinxemoji==0.2.0
pandas==1.4.1
pandoc>=2.1
matplotlib==3.5.1
